requirements.txt
flask
flask-socketio
flask-pymongo
pymongo
weasyprint
python-dotenv
eventlet
